# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Reading Tracker",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/books/{book_id}/engagements": {
      "get": {
        "tags": [
          "books"
        ],
        "summary": "List Book Engagements",
        "operationId": "books-list_book_engagements",
        "parameters": [
          {
            "name": "book_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Book Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EngagementRead"
                  },
                  "title": "Response Books-List Book Engagements"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/engagements": {
      "post": {
        "tags": [
          "engagements"
        ],
        "summary": "Create Engagement",
        "operationId": "engagements-create_engagement",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EngagementCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EngagementRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "engagements"
        ],
        "summary": "List Engagements",
        "operationId": "engagements-list_engagements",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/ReadingStatus"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EngagementRead"
                  },
                  "title": "Response Engagements-List Engagements"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/engagements/{engagement_id}": {
      "patch": {
        "tags": [
          "engagements"
        ],
        "summary": "Update Engagement Status",
        "operationId": "engagements-update_engagement_status",
        "parameters": [
          {
            "name": "engagement_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Engagement Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EngagementStatusUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EngagementRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "engagements"
        ],
        "summary": "Get Engagement",
        "operationId": "engagements-get_engagement",
        "parameters": [
          {
            "name": "engagement_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Engagement Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EngagementRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "engagements"
        ],
        "summary": "Delete Engagement",
        "operationId": "engagements-delete_engagement",
        "parameters": [
          {
            "name": "engagement_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Engagement Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/engagements/{engagement_id}/dates": {
      "patch": {
        "tags": [
          "engagements"
        ],
        "summary": "Update Engagement Dates",
        "operationId": "engagements-update_engagement_dates",
        "parameters": [
          {
            "name": "engagement_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Engagement Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EngagementDatesUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EngagementRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/engagements/{engagement_id}/length": {
      "patch": {
        "tags": [
          "engagements"
        ],
        "summary": "Update Engagement Length",
        "operationId": "engagements-update_engagement_length",
        "parameters": [
          {
            "name": "engagement_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Engagement Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EngagementLengthUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EngagementRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/engagements/{engagement_id}/editions": {
      "post": {
        "tags": [
          "engagements"
        ],
        "summary": "Create Binding",
        "operationId": "engagements-create_binding",
        "parameters": [
          {
            "name": "engagement_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Engagement Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EngagementEditionCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EngagementEditionRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "engagements"
        ],
        "summary": "List Bindings",
        "operationId": "engagements-list_bindings",
        "parameters": [
          {
            "name": "engagement_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Engagement Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EngagementEditionRead"
                  },
                  "title": "Response Engagements-List Bindings"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/engagements/{engagement_id}/editions/{edition_id}": {
      "delete": {
        "tags": [
          "engagements"
        ],
        "summary": "Delete Binding",
        "operationId": "engagements-delete_binding",
        "parameters": [
          {
            "name": "engagement_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Engagement Id"
            }
          },
          {
            "name": "edition_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Edition Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AuthorRead": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name"
        ],
        "title": "AuthorRead"
      },
...
      "BookRead": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "authors": {
            "items": {
              "$ref": "#/components/schemas/AuthorRead"
            },
            "type": "array",
            "title": "Authors"
          },
          "google_books_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Google Books Id"
          },
          "default_cover_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Default Cover Url"
          },
          "default_page_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Default Page Count"
          },
          "default_audio_minutes": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Default Audio Minutes"
          },
          "original_language": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Original Language"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "genres": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Genres"
          },
          "publication_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Publication Date"
          },
          "publication_date_precision": {
            "$ref": "#/components/schemas/DatePrecision"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "authors",
          "google_books_id",
          "default_cover_url",
          "default_page_count",
          "default_audio_minutes",
          "original_language",
          "description",
          "genres",
          "publication_date",
          "publication_date_precision",
          "created_at",
          "updated_at"
        ],
        "title": "BookRead"
      },
...
      "DatePrecision": {
        "type": "string",
        "enum": [
          "day",
          "month",
          "year"
        ],
        "title": "DatePrecision"
      },
...
      "EngagementCreate": {
        "properties": {
          "book_id": {
            "type": "string",
            "format": "uuid",
            "title": "Book Id"
          },
          "edition_format": {
            "$ref": "#/components/schemas/Format"
          },
          "status": {
            "type": "string",
            "enum": [
              "reading",
              "finished",
              "dnf"
            ],
            "title": "Status",
            "default": "reading"
          },
          "edition_length": {
            "anyOf": [
              {
                "type": "integer",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Edition Length"
          },
          "length_override": {
            "anyOf": [
              {
                "type": "integer",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Length Override"
          },
          "started_on": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started On"
          },
          "finished_on": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished On"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "book_id",
          "edition_format"
        ],
        "title": "EngagementCreate",
        "description": "`finished_on` is the date the read ended, whichever way it ended: it lands in\n`abandoned_on` when the status is dnf."
      },
      "EngagementDatesUpdate": {
        "properties": {
          "started_on": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started On"
          },
          "finished_on": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished On"
          },
          "abandoned_on": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Abandoned On"
          }
        },
        "type": "object",
        "title": "EngagementDatesUpdate",
        "description": "Corrects dates a read already has. Ending a read is the status endpoint's job:\n`finished_on` here edits a finished read, `abandoned_on` a dnf one."
      },
      "EngagementEditionCreate": {
        "properties": {
          "edition_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Edition Id"
          },
          "edition_format": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Format"
              },
              {
                "type": "null"
              }
            ]
          },
          "origin_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Origin Id"
          },
          "length_override": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Length Override"
          },
          "edition_length": {
            "anyOf": [
              {
                "type": "integer",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Edition Length"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "EngagementEditionCreate"
      },
      "EngagementEditionRead": {
        "properties": {
          "edition": {
            "$ref": "#/components/schemas/EditionRead"
          },
          "origin_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Origin Id"
          },
          "length_override": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Length Override"
          }
        },
        "type": "object",
        "required": [
          "edition",
          "origin_id",
          "length_override"
        ],
        "title": "EngagementEditionRead"
      },
      "EngagementLengthUpdate": {
        "properties": {
          "length_pages": {
            "anyOf": [
              {
                "type": "integer",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Length Pages"
          },
          "length_minutes": {
            "anyOf": [
              {
                "type": "integer",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Length Minutes"
          }
        },
        "type": "object",
        "title": "EngagementLengthUpdate",
        "description": "Mirrors the read side's length_pages / length_minutes pair. Exactly one, because\nthe unit is what tells the service which binding to correct."
      },
      "EngagementRead": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "book": {
            "$ref": "#/components/schemas/BookRead"
          },
          "formats": {
            "items": {
              "$ref": "#/components/schemas/Format"
            },
            "type": "array",
            "title": "Formats"
          },
          "cover_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cover Url"
          },
          "status": {
            "$ref": "#/components/schemas/ReadingStatus"
          },
          "started_on": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started On"
          },
          "finished_on": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished On"
          },
          "abandoned_on": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Abandoned On"
          },
          "resume_from_page": {
            "type": "integer",
            "title": "Resume From Page"
          },
          "resume_from_minute": {
            "type": "integer",
            "title": "Resume From Minute"
          },
          "frontier_page": {
            "type": "integer",
            "title": "Frontier Page"
          },
          "frontier_minute": {
            "type": "integer",
            "title": "Frontier Minute"
          },
          "resume_unit": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/LogUnit"
              },
              {
                "type": "null"
              }
            ]
          },
          "length_pages": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Length Pages"
          },
          "length_minutes": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Length Minutes"
          },
          "completion_pct": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completion Pct"
          },
          "review": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ReviewRead"
              },
              {
                "type": "null"
              }
            ]
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "book",
          "formats",
          "cover_url",
          "status",
          "started_on",
          "finished_on",
          "abandoned_on",
          "resume_from_page",
          "resume_from_minute",
          "frontier_page",
          "frontier_minute",
          "resume_unit",
          "length_pages",
          "length_minutes",
          "completion_pct",
          "review",
          "created_at",
          "updated_at"
        ],
        "title": "EngagementRead"
      },
      "EngagementStatusUpdate": {
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "reading",
              "finished",
              "dnf"
            ],
            "title": "Status"
          },
          "effective_on": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Effective On"
          },
          "unit": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/LogUnit"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "status"
        ],
        "title": "EngagementStatusUpdate",
        "description": "`unit` picks the ruler the closing log is written on when finishing a read that\nhas been going in more than one. Defaults to the one the read is already on."
      },
      "Format": {
        "type": "string",
        "enum": [
          "print",
          "digital",
          "audio"
        ],
        "title": "Format"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "LogUnit": {
        "type": "string",
        "enum": [
          "pages",
          "minutes"
        ],
        "title": "LogUnit"
      },
...
      "ReadingStatus": {
        "type": "string",
        "enum": [
          "interested",
          "tbr",
          "reading",
          "finished",
          "paused",
          "dnf"
        ],
        "title": "ReadingStatus"
      },
...
    }
  }
}